+++ scripts/test/SANS/user_file/CMakeLists.txt
@@ -5,7 +5,7 @@
 set ( TEST_PY_FILES
   user_file_parser_test.py
   user_file_reader_test.py
-  user_file_state_director_test.py
+  state_director_test.py
   )
 
 check_tests_valid ( ${CMAKE_CURRENT_SOURCE_DIR} ${TEST_PY_FILES} )
